Guarantee backfill for channels added at user reload during continuous changes feed
Under heavy access load, it's possible for the _changes feed to send the document that triggered an access grant before being notified that the user record has been updated and should be reloaded.  In this scenario backfill of the channel was being skipped.

This change forces a backfill of any new channels when the user gets reloaded - even if the seqAddedAt for that channel is earlier than the current since value.  This has the potential for duplicate entries being sent on the _changes feed if the client terminates mid-backfill, but these will get handled by revs diff.